+++ src/utils/storage.ts
@@ -1,6 +1,5 @@
 import { Release } from '../app/release.js';
 import { log, logError } from './console';
-import { generateKeyForUrl, generateKeysFromUrls } from './key-generator';
 import {
   getArrLastElement,
   getOwnProperty,
@@ -13,89 +12,34 @@ import { validate as isUUID } from 'uuid';
 
 const storage = chrome.storage.local;
 
+export type uuid = string;
+
 interface StorageData {
   [key: string]: any;
 }
 
-interface History extends Array<string> {}
+interface History extends Array<Date> {}
 
 export interface HistoryData {
-  [key: string]: History;
+  [key: uuid]: History;
 }
 
 interface ReleaseMap {
-  [key: string]: Release;
+  [key: uuid]: Release;
 }
 
 export interface VisitedDate {
-  uuid: string;
+  uuid: uuid;
   date: Date;
 }
-
-type ReleaseCallback = (release: Release) => void;
-type ReleasesCallback = (releases: Release[]) => void;
 
 export function logStorageData() {
   storage.get(null).then((data) => log('Storage data', data));
 }
 
-/**
- * Finds all releases from storage and release history and executes a callback function with the releases.
- */
-export function findAllReleases(): Promise<Release[]> {
+export function getAllReleases(): Promise<Release[]> {
   return storage.get(null).then((storageData: StorageData) => {
-    const releases: Release[] = [];
-
-    for (const key in storageData) {
-      if (
-        !hasOwnProperty(storageData, key) ||
-        !isObject(storageData[key]) ||
-        !isUUID(key)
-      ) {
-        continue;
-      }
-
-      try {
-        releases.push(Release.fromStorageObject(storageData[key]));
-      } catch (error) {
-        continue;
-      }
-    }
-
-    return releases;
-  });
-}
-
-/**
- * Finds a release by URL and executes the appropriate callback.
- * @param url - The URL to find the release for.
- * @param onFind - The callback function to call if the release is found.
- * @param onMissing - The callback function to call if the release is not found.
- */
-export function findReleaseByUrl(
-  url: string,
-  onFind?: ReleaseCallback,
-  onMissing?: (key: string) => void
-): void {
-  const key = generateKeyForUrl(url);
-
-  storage.get([key], (result: StorageData) => {
-    if (isObject(result[key])) {
-      try {
-        const release = Release.fromStorageObject(result[key]);
-        if (onFind && isFunction(onFind)) {
-          onFind(release);
-        }
-      } catch (error) {
-        log('Broken storage data for release', result[key]);
-        clearStorageByKey(key);
-      }
-    } else {
-      log('Release data is missing', key);
-      if (onMissing && isFunction(onMissing)) {
-        onMissing(key);
-      }
-    }
+    return storageDataToReleases(storageData);
   });
 }
 
@@ -114,62 +58,49 @@ export function getReleaseByUuid(uuid: string): Promise<Release | undefined> {
   });
 }
 
-export function getReleasesByUuids(uuids: string[]): Promise<ReleaseMap> {
+export function getReleaseMapByUuids(uuids: string[]): Promise<ReleaseMap> {
   return storage.get(uuids).then((storageData: StorageData) => {
-    const releases = getReleasesFromStorageData(storageData);
-    return createReleaseMapFromReleases(releases);
+    return storageDataToReleaseMap(storageData);
   });
 }
 
-const createReleaseMapFromReleases = (releases: Release[]) => {
-  const releaseMap: ReleaseMap = releases.reduce((acc, release) => {
-    acc[release.uuid] = release;
-    return acc;
-  }, {} as ReleaseMap);
-
-  return releaseMap;
-};
-
-const getReleasesFromStorageData = (storageData: StorageData) => {
-  const releases: Release[] = Object.values(storageData)
-    .map((obj: any) => {
-      try {
-        return Release.fromStorageObject(obj);
-      } catch (error) {
-        log('Broken storage object.', JSON.stringify(error), obj);
-        return null;
-      }
-    })
-    .filter((obj: Release | null): obj is Release => obj instanceof Release);
-
-  return releases;
-};
-
-/**
- * Finds releases by their URLs and executes a callback with the found releases.
- */
-export function findReleasesByUrls(
-  urls: string[],
-  onFind?: ReleasesCallback
-): void {
-  const keys = generateKeysFromUrls(urls);
-
-  storage.get(keys, (result: StorageData) => {
-    const releases: Release[] = Object.values(result)
-      .map((obj: any) => {
-        try {
-          return Release.fromStorageObject(obj);
-        } catch (error) {
-          log('Broken storage object.', JSON.stringify(error), obj);
-          return null;
-        }
-      })
-      .filter((obj: Release | null): obj is Release => obj instanceof Release);
-
-    if (onFind && isFunction(onFind)) {
-      onFind(releases);
-    }
+export function getReleasesByUuids(uuids: string[]): Promise<Release[]> {
+  return storage.get(uuids).then((storageData: StorageData) => {
+    return storageDataToReleases(storageData);
   });
+}
+
+function releaseMapToReleases(releaseMap: ReleaseMap): Release[] {
+  return Object.values(releaseMap);
+}
+
+function storageDataToReleases(storageData: StorageData): Release[] {
+  return releaseMapToReleases(storageDataToReleaseMap(storageData));
+}
+
+function storageDataToReleaseMap(storageData: StorageData): ReleaseMap {
+  const releasesMap: ReleaseMap = {};
+
+  for (const key in storageData) {
+    if (
+      !hasOwnProperty(storageData, key) ||
+      !isObject(storageData[key]) ||
+      !isUUID(key)
+    ) {
+      continue;
+    }
+
+    const obj = storageData[key];
+
+    try {
+      releasesMap[key] = Release.fromStorageObject(obj);
+    } catch (error) {
+      log('Broken release storage object.', JSON.stringify(error), obj);
+      continue;
+    }
+  }
+
+  return releasesMap;
 }
 
 /**
@@ -184,17 +115,32 @@ export function saveRelease(release: Release): void {
 
 export function addReleaseHistory(release: Release): void {
   const uuid = release.uuid;
-  getHistoryByUuid(uuid).then((history) => {
-    history.push(new Date().toISOString());
+  getHistoryByUuid(uuid).then((historyData) => {
+    const history = getOwnProperty(historyData, uuid, []);
+    history.push(new Date());
     setHistoryByUuid(uuid, history).then(() => {
       log('New history added to release', uuid);
     });
   });
 }
 
-export function getHistoryByUuid(uuid: string): Promise<History> {
-  const key = generateHistoryKey(uuid);
-  return storage.get([key]).then((result) => getOwnProperty(result, key, []));
+export function getHistoryByUuid(uuid: string): Promise<HistoryData> {
+  const historyKey = generateHistoryKey(uuid);
+  return storage.get([historyKey]).then((result) => {
+    const historyData: HistoryData = {};
+    historyData[uuid] = dateStringsToDateObjects(
+      getOwnProperty(result, historyKey, [])
+    );
+    return historyData;
+  });
+}
+
+function dateStringsToDateObjects(dates: string[]): Date[] {
+  return dates.map((dateString) => new Date(dateString));
+}
+
+function dateObjectsToDateStrings(dates: Date[]): string[] {
+  return dates.map((date) => date.toISOString());
 }
 
 export function setHistoryByUuid(
@@ -202,7 +148,7 @@ export function setHistoryByUuid(
   history: History
 ): Promise<void> {
   const key = generateHistoryKey(uuid);
-  return storage.set({ [key]: history });
+  return storage.set({ [key]: dateObjectsToDateStrings(history) });
 }
 
 export function getHistoryData(): Promise<HistoryData> {
@@ -213,7 +159,7 @@ export function getHistoryData(): Promise<HistoryData> {
       if (!key.startsWith(HISTORY_KEY_PREFIX)) continue;
       const uuid = key.slice(HISTORY_KEY_PREFIX.length);
       if (!isUUID(uuid)) continue;
-      historyData[uuid] = storageData[key];
+      historyData[uuid] = dateStringsToDateObjects(storageData[key]);
     }
 
     return historyData;