+++ CHANGELOG.md
@@ -1,6 +1,19 @@
 # Change log
 
-## 0.15.1 (2024-10-XX)
+## 0.16.0 (2024-10-08)
+
+### Features
+
+- core: Refactor rendering Release Card
+- core: A small refactor of getting data from the local storage
+- popup: Make popup wider
+- popup: Moved download CSV button and search buttons to the Release Card
+- popup: Load release data into the release card from the history tab and latest visited releases widget
+- popup: Output release info, credits, tracklist and history information on the release card
+
+### Fixed
+
+## 0.15.1 (2024-10-01)
 
 ### Features